Renamed build-release command to build-release-locally
