+++ scripts/release/build-release-locally-commands/add-build-info-json.js
