+++ scripts/release/build-release-locally-commands/confirm-automated-testing.js
